requirements.txt
pip==23.0
certifi==2022.12.7
requests==2.28.1
pydantic==1.10.4
idna==3.4
urllib3==1.26.13
setuptools==56.0.0
python-dotenv==0.21.0
python-telegram-bot==13.7
telegram==0.0.1
numpy==1.24.1
# bottle==0.12.23